Fix simplification branch verification
