+++ packages/backend/test/identity-recovery.test.mjs
@@ -1,3 +1,4 @@
+/* eslint-disable no-empty */
 import assert from 'node:assert/strict'
 import fs from 'node:fs'
 import os from 'node:os'
